explicitly pass in the config_entry in coordinator init
